Initial Commit
This is more of a mock-up right now. The code is all in one
file for now (index.html), and the timer itself is static, and
does not actually tick for each second yet.

The design works on Google Chrome and Firefox, however some of it
gets messed up on IE. Also, the script in the HTML file takes a
long time to finish in Firefox, and the browser shows a warning
message about it.