Added an opening form for the viewer's DOB
The new page now opens with a simple form asking for the viewer's
date of birth, rather than jumping straight into the diagram
itself. Some more Javascript has to be used to get and interpret
the given date of birth, as well as to replace the form div with
the description div and to make the rest of the page visible.