Added an animated error message if an invalid birthday is given
This animated error message is created in the HTML but is hidden
by setting the text color to black in stylesheet.css. However, if
the user clicks the go button without entering a valid birthday,
then the function showIvalidText() will be called and will attach
an animated class to the hidden error text to make it smoothly
appear.