Fixed minor problem with 'go' button
Since the button is actually from an <a> tag, it redirects the
user to another link when clicked (in this case, href is just '#').
However, there is one trick in javascript to keep the link from
changing the URL, which is to return 'false' for the onclick
function. This commit adds this, so the button no longer adds the
'#' to the end of the URL.