Added fade-in animation for main content
All of the birthdate UI stuff is now done! The webpage successfully
starts off by asking for the user's birthdate, and then smoothly
transitions to the main page and rectangle once a valid birthday
is given.

This fade-in animation is done with the same technique as the other
JS-invoked animations.